Change encryption lib
