+++ .docs/Examples/04-Encrypted-Storage.md
@@ -5,8 +5,10 @@ order: 4
 
 # Encrypt persisted data
 
-You can encrypt data that are persisted.  
-Both the key and the data are encrypted.
+You can encrypt data that are persisted.
+
+The key is hashed (with the [[GCMEncryption]], hashing is done with an encryption).  
+The data is encrypted.
 
 _**NOTE:** The encryption is not well adapted for cookie storage, as the data is encrypted, its size greatly increase, you will rapidly get to the allowed cookie size._
 
@@ -14,10 +16,10 @@ _**NOTE:** The encryption is not well adapted for cookie storage, as the data is
 
 ```html
 <script>
-    import { persist, createLocalStorage, createEncryptedStorage } from "@macfja/svelte-persistent-store"
+    import { persist, createLocalStorage, createEncryptedStorage, GCMEncryption } from "@macfja/svelte-persistent-store"
     import { writable } from "svelte/store"
 
-    const storage = createEncryptedStorage(createLocalStorage(), "5368566D597133743677397A24432646")
+    const storage = createEncryptedStorage(createLocalStorage(), new GCMEncryption("5368566D597133743677397A24432646"))
     const nickname = persist(writable("John"), storage, "myapp-nickname")
 </script>
 ```
