+++ package.json
@@ -21,8 +21,9 @@
     "dependencies": {
         "@macfja/serializer": "^1.1.1",
         "browser-cookies": "^1.2.0",
-        "cyrup": "^0.8.1",
-        "idb-keyval": "^5.1.3"
+        "idb-keyval": "^5.1.3",
+        "sjcl-codec-hex": "^1.0.0",
+        "sjcl-es": "^2.0.0"
     },
     "devDependencies": {
         "@rollup/plugin-commonjs": "^22.0.0",
